hackathonfinal-main/pyproject.toml
[project]
name = "version-4-multi-agent-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncclick>=8.1.8",
    "click>=8.1.8",
    "google-adk>=1.0.0",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "langchain-google-genai>=2.1.4",
    "langchain-mcp-adapters>=0.0.11",
    "langgraph>=0.4.3",
    "mcp>=1.8.0",
    "psutil>=7.0.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]
